LangChain/project/chatbot-app/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.7
redis==5.2.0
python-dotenv==1.0.1
